authentication and authrisation using jwt done

+++ controllers/api/v1/post_api.js
@@ -20,20 +20,21 @@ module.exports.index = async (req,res)=>{
 module.exports.destroy = async (req, res) => {
     try {
       let post = await Post.findById(req.params.id);
-    //   if (post.user == req.user.id) {
+      if (post.user == req.user.id) {
         post.remove();
         await Comment.deleteMany({ post: req.params.id });
   
-    
-  
-        
         return res.json(200,{
             message:"post and associated comments deleted succesfully"
         })
-    //   } else {
-    //     req.flash("error","you cannot delete this post!")
-    //     return res.redirect("back");
-    //   }
+    
+  
+        
+      } else {
+        return res.json(401,{
+          message:"you cannot delete this post"
+      })
+      }
     } catch (error) {
      
       return res.json(500,{
